+++ tests/js/list-tests.js
@@ -1470,11 +1470,75 @@ module.exports = {
             addresses[0]["street"] = "Strandvejen";
         });
 
+        // insert an extra address into Ib's list (add embedded object)
+        let ibs_addrs = owners[0].addresses;
+        realm.write(() => {
+            ibs_addrs.push({ street: "Njalsgade", city: "Islands Brygge" });
+        });
+
+        streets = ["Algade", "Njalsgade", "Skolevej", "Strandvejen"];
+        for (let i = 0; i < streets.length; i++) {
+            TestCase.assertEqual(addresses[i]["street"], streets[i]);
+        }
+
+        // remove the last of Ib's addresses
+        realm.write(() => {
+            ibs_addrs.pop();
+        });
         streets = ["Algade", "Skolevej", "Strandvejen"];
         for (let i = 0; i < streets.length; i++) {
             TestCase.assertEqual(addresses[i]["street"], streets[i]);
         }
 
+        realm.close();
+    },
+
+    testCreateNestedEmbeddedObjects: function() {
+        const realm = new Realm({schema: [schemas.ScoutDivisionSchema, schemas.ScoutGroupSchema, schemas.ScoutBranchSchema]});
+
+        realm.write(() => {
+            realm.create(schemas.ScoutDivisionSchema.name, {
+                name: "Oeresund Division", 
+                groups: [
+                    {
+                        name: "RungstedSpejderne",
+                        branches: [ { name: "Micro" }, { name: "Mini" }, { name: "Junior" }, {name: "Trop" } ]
+                    },
+                    {
+                        name: "Bent Byg",
+                        branches: [ { name: "Mini" }, { name: "Junior" }, { name: "Trop" }, { name: "Klan" } ]
+                    }
+                ]
+            });
+            realm.create(schemas.ScoutDivisionSchema.name, {
+                name: "Bernstorff Division",
+                groups: [
+                    {
+                        name: 'HellerupSpejderne',
+                        branches: [ { name: 'Mini' }, { name: 'Flok' }, { name: 'Klan' } ]
+                    }
+                ]
+            });
+        });
+
+        let divisions = realm.objects(schemas.ScoutDivisionSchema.name).sorted("name");
+        let groups = realm.objects(schemas.ScoutGroupSchema.name).sorted("name");
+        TestCase.assertEqual(divisions.length, 2);
+        TestCase.assertEqual(groups.length, 3);
+
+        let bernstorff_groups = divisions[0].groups;
+        TestCase.assertEqual(bernstorff_groups.length, 1);
+
+        // add a Group to Bernstorff Division
+        realm.write(() => {
+            bernstorff_groups.push({
+                name: "1. Ordrup",
+                branches: [ { name: 'FamilieSpejd' }, { name: 'Mikro' }, { name: 'Ulve' }, { name: 'Hvalpe' }, { name: 'Trop' }, { name: 'Klan' } ]
+            });
+        });
+
+        // check that we have successfully added a Group
+        TestCase.assertEqual(groups.length, 4);
         realm.close();
     },
 
@@ -1506,8 +1570,11 @@ module.exports = {
 
         realm.write(() => {
             TestCase.assertThrows(() => {
-                ib.addresses.push({ street: "Njalsgade", city: "Islands Brygge" });
+                realm.create(schemas.AddressSchema.name, { street: "Njalsgade", city: "Islands Brygge" });
             });
+
+            ib.addresses.push({ street: "Njalsgade", city: "Islands Brygge" });
+            TestCase.assertEqual(3, ib.addresses.length);
         });
 
         realm.close();